Move the matchers in CollectionIncompatibleType out for reuse, as well as the types in UndefinedEquals.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=308253156
